requirements.txt
# Dependencies
selenium>=4.0.0
pyyaml>=6.0
